@@ -12,49 +12,66 @@ dependencies:
     sdk: flutter
   persistent_bottom_nav_bar: ^5.0.2
   ionicons: ^0.2.2
-  iconsax: ^0.0.8
+  # iconsax: ^0.0.8
   enefty_icons: ^1.0.4
-  iconly: ^1.0.1
+  # iconly: ^1.0.1
   qr_code_scanner: ^1.0.0
   simple_animations: ^5.0.2
   email_validator: ^2.1.17
   introduction_screen: ^3.1.11
-  google_fonts: ^5.1.0
-  contained_tab_bar_view: ^0.8.0
+  # google_fonts: ^5.1.0
+  # contained_tab_bar_view: ^0.8.0
   intl: ^0.18.1
-  freezed_annotation: ^2.4.1
-  build_runner: ^2.4.6
-  freezed: ^2.4.2
+  # freezed_annotation: ^2.4.1
+  # build_runner: ^2.4.6
+  # freezed: ^2.4.2
   http: ^1.1.0
-  dio: ^5.3.2
+  dio: ^5.4.1
   device_info_plus: ^4.1.2
   animated_toggle_switch: ^0.8.0
   flutter_switch: ^0.3.2
   get_storage: ^2.1.1
-  qr_flutter: ^4.0.0
-  mobile_scanner: ^2.1.0
+  qr_flutter: ^4.1.0
+  # mobile_scanner: ^2.1.0
   shimmer: ^3.0.0
   lottie: ^2.6.0
-  google_sign_in: ^6.1.5
-  googleapis: ^11.4.0
-  syncfusion_flutter_calendar: ^23.1.40
+  # google_sign_in: ^6.1.5
+  # googleapis: ^11.4.0
+  # syncfusion_flutter_calendar: ^23.1.40
   image_picker: ^1.0.4
   scan: ^1.6.0
-  camera: ^0.10.5+5
+  # camera: ^0.10.5+5
   image_cropper: ^5.0.0
-  location: ^5.0.3
+  # location: ^5.0.3
   geolocator: ^10.1.0
   geocoding: ^2.1.1
+  scrollable_positioned_list: ^0.3.8
+  percent_indicator: ^4.2.3
+  day_night_time_picker: ^1.3.0+1
+  share: ^2.0.4
+  path_provider: ^2.1.2
+  # provider: ^6.1.2
+  flutter_slidable: ^3.0.1
+  awesome_dialog: ^3.2.0
+  horizontal_data_table: ^4.3.1
+  permission_handler: ^11.0.1
+  dropdown_button2: ^2.3.9
+  dotted_border: ^2.1.0
+  table_calendar: ^3.0.9
 
 dev_dependencies:
   flutter_lints: ^2.0.0
   flutter_test:
     sdk: flutter
-  json_serializable: ^6.7.1
+  # json_serializable: ^6.7.1
 
 flutter:
   assets:
     - assets/images/
+    - assets/images/FAQ/tutorial-menambahkan-dosen/
+    - assets/images/FAQ/tutorial-membuat-pertemuan/
+    - assets/images/FAQ/tutorial-menghapus-pertemuan/
+    - assets/images/FAQ/tutorial-membuat-qrcode/
     - assets/icons/
     - assets/fonts/
     - assets/lottie/
@@ -77,4 +94,3 @@ flutter:
     - family: Signika
       fonts:
         - asset: assets/fonts/Signika-Light.ttf
-
