@@ -1,6 +1,7 @@
+import 'package:absensi_mahardika/app/utils/color.dart';
+import 'package:absensi_mahardika/app/utils/style.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:syncfusion_flutter_calendar/calendar.dart' as calendar;
 import '../controllers/calendar_controller.dart';
 
 class CalendarView extends GetView<CalendarController> {
@@ -8,38 +9,106 @@ class CalendarView extends GetView<CalendarController> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
+      backgroundColor: AppColor.bgColor,
       appBar: AppBar(
-        title: Text(controller.data),
+        title: const Text("calendar"),
         centerTitle: true,
       ),
-      body: FutureBuilder(
-        future: controller.getGoogleEventsData(),
-        builder: (BuildContext context, AsyncSnapshot snapshot) {
-          print(controller.currentUser);
-
-          return Stack(
-            children: [
-              calendar.SfCalendar(
-                view: calendar.CalendarView.month,
-                cellBorderColor: Colors.transparent,
-                initialDisplayDate: DateTime.now(),
-                dataSource: GoogleDataSource(events: snapshot.data),
-                monthViewSettings: const calendar.MonthViewSettings(
-                  appointmentDisplayMode:
-                      calendar.MonthAppointmentDisplayMode.indicator,
-
-                  showAgenda: true,
-                ),
-                
+      body: Padding(
+        padding: const EdgeInsets.symmetric(horizontal: 15.0),
+        child: ListView.builder(
+          itemCount: 3,
+          physics: const ScrollPhysics(),
+          itemBuilder: (BuildContext context, int index) {
+            return Container(
+              margin: const EdgeInsets.only(
+                bottom: 10.0,
               ),
-              snapshot.data != null
-                  ? Container()
-                  : const Center(
-                      child: CircularProgressIndicator(),
-                    )
-            ],
-          );
-        },
+              padding: const EdgeInsets.all(12.0),
+              height: Get.height / 7.5,
+              decoration: const BoxDecoration(
+                color: Colors.white,
+                borderRadius: BorderRadius.all(
+                  Radius.circular(12.0),
+                ),
+              ),
+              child: Row(
+                children: [
+                  Container(
+                    height: Get.height / 9,
+                    padding: const EdgeInsets.all(5.0),
+                    decoration: const BoxDecoration(),
+                    child: const Image(
+                      image: AssetImage('assets/icons/education1.png'),
+                    ),
+                  ),
+                  const SizedBox(
+                    width: 10.0,
+                  ),
+                  Expanded(
+                    child: Column(
+                      children: [
+                        Row(
+                          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                          children: [
+                            Text(
+                              "Struktur Data",
+                              style: Style.Header1.copyWith(
+                                  color: AppColor.blackSoftColor),
+                            ),
+                            Container(
+                              child: const Row(
+                                mainAxisAlignment:
+                                    MainAxisAlignment.spaceBetween,
+                                children: [
+                                  Icon(
+                                    Icons.edit_note_sharp,
+                                    size: 24.0,
+                                    color: Colors.grey,
+                                  ),
+                                  SizedBox(
+                                    width: 10.0,
+                                  ),
+                                  Icon(
+                                    Icons.more_horiz_outlined,
+                                    size: 24.0,
+                                    color: AppColor.bluePrimary,
+                                  ),
+                                ],
+                              ),
+                            ),
+                          ],
+                        ),
+                        const Spacer(),
+                        Row(
+                          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                          children: [
+                            Text(
+                              "21 April 2024",
+                              style: Style.Subtitle1.copyWith(
+                                  color:
+                                      AppColor.blackSoftColor.withOpacity(0.5)),
+                            ),
+                            Container(
+                              height: 20,
+                              width: 50,
+                              decoration: BoxDecoration(
+                                borderRadius: const BorderRadius.all(
+                                  Radius.circular(4.0),
+                                ),
+                                color: Colors.green[100]!.withOpacity(0.8),
+                              ),
+                            ),
+                          ],
+                        ),
+                      ],
+                    ),
+                  ),
+                ],
+              ),
+            );
+          },
+        ),
       ),
     );
   }
