@@ -1,5 +1,7 @@
 import 'dart:async';
 import 'dart:convert';
+import 'package:absensi_mahardika/app/modules/Khs/views/khs_view.dart';
+import 'package:absensi_mahardika/app/modules/krs/views/krs_view.dart';
 import 'package:absensi_mahardika/app/utils/color.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
@@ -9,8 +11,11 @@ import '../../../data/totalAbsen_model.dart';
 import '../../../utils/network.dart';
 import "package:http/http.dart" as http;
 
+import '../../perizinan/views/perizinan_view.dart';
+
 class HomeController extends GetxController
     with GetSingleTickerProviderStateMixin {
+  String hours = '';
   int hour = 0;
   int minute = 0;
   String time = '';
@@ -25,6 +30,54 @@ class HomeController extends GetxController
   var dataUserLocal = box.read('dataUser');
   var days = DateFormat('EEEE, d MMM yyyy', "id_ID").format(DateTime.now());
 
+  List<Map<String, dynamic>> listFitur = [
+    {
+      'image': "assets/icons/krs.png",
+      'name': "Krs".tr,
+      'subtitle': "Kartu Rencana Studi",
+      'onTap': () {
+        Get.to(
+          () => const KrsView(),
+          transition: Transition.rightToLeft,
+        );
+      }
+    },
+    {
+      'image': "assets/icons/khs.png",
+      'name': "Khs".tr,
+      'subtitle': "Kartu Hasil Studi",
+      'onTap': () {
+        Get.to(
+          () => const KhsView(),
+          transition: Transition.rightToLeft,
+        );
+      }
+    },
+    {
+      'image': "assets/icons/transkip.png",
+      'name': "Transkrip".tr,
+      'subtitle': 'Download Transkrip Nilai',
+      'onTap': () {
+        Get.to(() => const KrsView(), transition: Transition.rightToLeft);
+      }
+    },
+    {
+      'image': "assets/icons/permit.png",
+      'name': "Perizinan".tr,
+      'subtitle': 'Izin Perkuliahan',
+      'onTap': () {
+        Get.to(() => const PerizinanView(), transition: Transition.rightToLeft);
+      }
+    },
+    {
+      'image': "assets/icons/todo.png",
+      'name': "ToDoListKa".tr,
+      'subtitle': 'Ayo Atur Tugas Kuliahmu',
+      'onTap': () {
+        Get.to(() => const KrsView(), transition: Transition.rightToLeft);
+      }
+    },
+  ];
   void changeTheme() {
     Get.changeTheme(Get.isDarkMode ? Themes.light : Themes.dark);
   }
@@ -34,43 +87,36 @@ class HomeController extends GetxController
       await Future.delayed(const Duration(seconds: 1));
       time = DateFormat('yyyy-MM-dd').format(DateTime.now());
       hour = int.parse(DateFormat('HH').format(DateTime.now()));
+      hours = DateFormat('HH:mm').format(DateTime.now());
       minute = int.parse(DateFormat('mm').format(DateTime.now()));
       yield DateTime.now();
     }
   }
 
+  // Stream<totalAbsenModel> streamAbsen(String npm) async* {
+  //   var url = Uri.http(baseUrl, 'api/absensi', {'npm': npm});
 
-  Stream<totalAbsenModel> streamAbsen(String npm) async* {
-    var url = Uri.http(baseUrl, 'api/absensi', {'npm': npm});
+  //   yield* Stream.periodic(const Duration(seconds: 15), (_) {
+  //     return http.get(
+  //       url,
+  //       headers: {
+  //         'Accept': 'application/json',
+  //         'Authorization': 'Bearer $token',
+  //       },
+  //     );
+  //   }).asyncMap((event) async {
+  //     var data = await event;
 
-    yield* Stream.periodic(const Duration(seconds: 15), (_) {
-      return http.get(
-        url,
-        headers: {
-          'Accept': 'application/json',
-          'Authorization': 'Bearer $token',
-        },
-      );
-    }).asyncMap((event) async {
-      var data = await event;
+  //     var result = jsonDecode(data.body);
 
-      var result = jsonDecode(data.body);
-
-      totalAbsen.value = totalAbsenModel.fromJson(result);
-      print(token);
-      return totalAbsen.value;
-    });
-  }
-
+  //     totalAbsen.value = totalAbsenModel.fromJson(result);
+  //     print(token);
+  //     return totalAbsen.value;
+  //   });
+  // }
 
   @override
   void onInit() {
-    // TODO: implement onInit
-    // newsStream.listen((event) async {
-    //   var data = await event;
-    //   totalAbsen.value = totalAbsenModel.fromJson(data);
-    //   print(token);
-    // });
     GetStorage.init();
     super.onInit();
   }
@@ -81,95 +127,3 @@ class HomeController extends GetxController
     super.onClose();
   }
 }
-
-// class LifeCycleController extends SuperController {
-//   var totalAbsen = totalAbsenModel().obs;
-//   late StreamController<totalAbsenModel> streamController;
-//   late Stream timerStream;
-//   final Stream newsStream =
-//       Stream.periodic(const Duration(seconds: 5), (_) async {
-//     try {
-//       var url = Uri.http(
-//           baseUrl, 'api/absensi', {'npm': dataUser.value.account!.npm});
-//       http.Response response = await http.get(
-//         url,
-//         headers: {
-//           'Accept': 'application/json',
-//           'Authorization': 'Bearer $token',
-//         },
-//       );
-
-//       var result = jsonDecode(response.body);
-//       return result;
-//     } catch (e) {
-//       print(e);
-//     }
-//   });
-//   StreamSubscription? sub;
-
-//   Stream<totalAbsenModel> streamAbsen(String npm) async* {
-//     newsStream.asyncMap((event) async {
-//       var data = await event;
-//       totalAbsen.value = totalAbsenModel.fromJson(data);
-
-//       // print(token);
-//       return totalAbsen.value;
-//     });
-//   }
-
-//   @override
-//   void onDetached() {
-//     print('1');
-//   }
-
-//   @override
-//   void onInactive() {
-//     print('2');
-//   }
-
-//   @override
-//   void onPaused() {
-//     sub!.cancel();
-//     print('3');
-//   }
-
-//   @override
-//   void onResumed() {
-//     print('4');
-//   }
-
-//   @override
-//   void onInit() {
-//     // TODO: implement onInit
-//     timerStream = Stream.periodic(const Duration(seconds: 5), (_) async {
-//       try {
-//         var url = Uri.http(
-//             baseUrl, 'api/absensi', {'npm': dataUser.value.account!.npm});
-//         http.Response response = await http.get(
-//           url,
-//           headers: {
-//             'Accept': 'application/json',
-//             'Authorization': 'Bearer $token',
-//           },
-//         );
-
-//         var result = jsonDecode(response.body);
-//         return result;
-//       } catch (e) {
-//         print(e);
-//       }
-//     });
-
-//     streamController = StreamController.broadcast();
-//     var timerSubscription = timerStream.listen((event) {
-//       print(event);
-//       streamController.add(event);
-//     });
-//     // sub = newsStream.listen((event) async {
-//     //   var data = await event;
-//     //   totalAbsen.value = totalAbsenModel.fromJson(data);
-//     //   print(token);
-//     // });
-//     super.onInit();
-//   }
-// }
