@@ -1,6 +1,7 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
 // ignore_for_file: camel_case_types, prefer_typing_uninitialized_variables
 
+import 'package:absensi_mahardika/app/utils/color.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
@@ -21,10 +22,11 @@ class _BgContainerState extends State<BgContainer> {
     return Container(
       width: Get.width,
       margin: const EdgeInsets.only(top: 60),
-      padding: const EdgeInsets.only(top: 100, left: 25, right: 25),
-      decoration: BoxDecoration(
-        color: context.theme.primaryColor,
-        borderRadius: const BorderRadius.all(
+      padding: const EdgeInsets.only(top: 100, left: 15, right: 15),
+      decoration: const BoxDecoration(
+        // color: context.theme.primaryColor,
+        color: AppColor.bgColor,
+        borderRadius: BorderRadius.all(
           Radius.circular(15.0),
         ),
       ),
