@@ -1,186 +1,186 @@
-// coverage:ignore-file
-// GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: type=lint
-// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
+// // coverage:ignore-file
+// // GENERATED CODE - DO NOT MODIFY BY HAND
+// // ignore_for_file: type=lint
+// // ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
 
-part of 'device_info_model.dart';
+// part of 'device_info_model.dart';
 
-// **************************************************************************
-// FreezedGenerator
-// **************************************************************************
+// // **************************************************************************
+// // FreezedGenerator
+// // **************************************************************************
 
-T _$identity<T>(T value) => value;
+// T _$identity<T>(T value) => value;
 
-final _privateConstructorUsedError = UnsupportedError(
-    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+// final _privateConstructorUsedError = UnsupportedError(
+//     'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
 
-DeviceInfoModel _$DeviceInfoModelFromJson(Map<String, dynamic> json) {
-  return _DeviceInfoModel.fromJson(json);
-}
+// DeviceInfoModel _$DeviceInfoModelFromJson(Map<String, dynamic> json) {
+//   return _DeviceInfoModel.fromJson(json);
+// }
 
-/// @nodoc
-mixin _$DeviceInfoModel {
-  String? get model => throw _privateConstructorUsedError;
-  String? get id => throw _privateConstructorUsedError;
-  String? get version => throw _privateConstructorUsedError;
+// /// @nodoc
+// mixin _$DeviceInfoModel {
+//   String? get model => throw _privateConstructorUsedError;
+//   String? get id => throw _privateConstructorUsedError;
+//   String? get version => throw _privateConstructorUsedError;
 
-  Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
-  @JsonKey(ignore: true)
-  $DeviceInfoModelCopyWith<DeviceInfoModel> get copyWith =>
-      throw _privateConstructorUsedError;
-}
+//   Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
+//   @JsonKey(ignore: true)
+//   $DeviceInfoModelCopyWith<DeviceInfoModel> get copyWith =>
+//       throw _privateConstructorUsedError;
+// }
 
-/// @nodoc
-abstract class $DeviceInfoModelCopyWith<$Res> {
-  factory $DeviceInfoModelCopyWith(
-          DeviceInfoModel value, $Res Function(DeviceInfoModel) then) =
-      _$DeviceInfoModelCopyWithImpl<$Res, DeviceInfoModel>;
-  @useResult
-  $Res call({String? model, String? id, String? version});
-}
+// /// @nodoc
+// abstract class $DeviceInfoModelCopyWith<$Res> {
+//   factory $DeviceInfoModelCopyWith(
+//           DeviceInfoModel value, $Res Function(DeviceInfoModel) then) =
+//       _$DeviceInfoModelCopyWithImpl<$Res, DeviceInfoModel>;
+//   @useResult
+//   $Res call({String? model, String? id, String? version});
+// }
 
-/// @nodoc
-class _$DeviceInfoModelCopyWithImpl<$Res, $Val extends DeviceInfoModel>
-    implements $DeviceInfoModelCopyWith<$Res> {
-  _$DeviceInfoModelCopyWithImpl(this._value, this._then);
+// /// @nodoc
+// class _$DeviceInfoModelCopyWithImpl<$Res, $Val extends DeviceInfoModel>
+//     implements $DeviceInfoModelCopyWith<$Res> {
+//   _$DeviceInfoModelCopyWithImpl(this._value, this._then);
 
-  // ignore: unused_field
-  final $Val _value;
-  // ignore: unused_field
-  final $Res Function($Val) _then;
+//   // ignore: unused_field
+//   final $Val _value;
+//   // ignore: unused_field
+//   final $Res Function($Val) _then;
 
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? model = freezed,
-    Object? id = freezed,
-    Object? version = freezed,
-  }) {
-    return _then(_value.copyWith(
-      model: freezed == model
-          ? _value.model
-          : model // ignore: cast_nullable_to_non_nullable
-              as String?,
-      id: freezed == id
-          ? _value.id
-          : id // ignore: cast_nullable_to_non_nullable
-              as String?,
-      version: freezed == version
-          ? _value.version
-          : version // ignore: cast_nullable_to_non_nullable
-              as String?,
-    ) as $Val);
-  }
-}
+//   @pragma('vm:prefer-inline')
+//   @override
+//   $Res call({
+//     Object? model = freezed,
+//     Object? id = freezed,
+//     Object? version = freezed,
+//   }) {
+//     return _then(_value.copyWith(
+//       model: freezed == model
+//           ? _value.model
+//           : model // ignore: cast_nullable_to_non_nullable
+//               as String?,
+//       id: freezed == id
+//           ? _value.id
+//           : id // ignore: cast_nullable_to_non_nullable
+//               as String?,
+//       version: freezed == version
+//           ? _value.version
+//           : version // ignore: cast_nullable_to_non_nullable
+//               as String?,
+//     ) as $Val);
+//   }
+// }
 
-/// @nodoc
-abstract class _$$_DeviceInfoModelCopyWith<$Res>
-    implements $DeviceInfoModelCopyWith<$Res> {
-  factory _$$_DeviceInfoModelCopyWith(
-          _$_DeviceInfoModel value, $Res Function(_$_DeviceInfoModel) then) =
-      __$$_DeviceInfoModelCopyWithImpl<$Res>;
-  @override
-  @useResult
-  $Res call({String? model, String? id, String? version});
-}
+// /// @nodoc
+// abstract class _$$_DeviceInfoModelCopyWith<$Res>
+//     implements $DeviceInfoModelCopyWith<$Res> {
+//   factory _$$_DeviceInfoModelCopyWith(
+//           _$_DeviceInfoModel value, $Res Function(_$_DeviceInfoModel) then) =
+//       __$$_DeviceInfoModelCopyWithImpl<$Res>;
+//   @override
+//   @useResult
+//   $Res call({String? model, String? id, String? version});
+// }
 
-/// @nodoc
-class __$$_DeviceInfoModelCopyWithImpl<$Res>
-    extends _$DeviceInfoModelCopyWithImpl<$Res, _$_DeviceInfoModel>
-    implements _$$_DeviceInfoModelCopyWith<$Res> {
-  __$$_DeviceInfoModelCopyWithImpl(
-      _$_DeviceInfoModel _value, $Res Function(_$_DeviceInfoModel) _then)
-      : super(_value, _then);
+// /// @nodoc
+// class __$$_DeviceInfoModelCopyWithImpl<$Res>
+//     extends _$DeviceInfoModelCopyWithImpl<$Res, _$_DeviceInfoModel>
+//     implements _$$_DeviceInfoModelCopyWith<$Res> {
+//   __$$_DeviceInfoModelCopyWithImpl(
+//       _$_DeviceInfoModel _value, $Res Function(_$_DeviceInfoModel) _then)
+//       : super(_value, _then);
 
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? model = freezed,
-    Object? id = freezed,
-    Object? version = freezed,
-  }) {
-    return _then(_$_DeviceInfoModel(
-      model: freezed == model
-          ? _value.model
-          : model // ignore: cast_nullable_to_non_nullable
-              as String?,
-      id: freezed == id
-          ? _value.id
-          : id // ignore: cast_nullable_to_non_nullable
-              as String?,
-      version: freezed == version
-          ? _value.version
-          : version // ignore: cast_nullable_to_non_nullable
-              as String?,
-    ));
-  }
-}
+//   @pragma('vm:prefer-inline')
+//   @override
+//   $Res call({
+//     Object? model = freezed,
+//     Object? id = freezed,
+//     Object? version = freezed,
+//   }) {
+//     return _then(_$_DeviceInfoModel(
+//       model: freezed == model
+//           ? _value.model
+//           : model // ignore: cast_nullable_to_non_nullable
+//               as String?,
+//       id: freezed == id
+//           ? _value.id
+//           : id // ignore: cast_nullable_to_non_nullable
+//               as String?,
+//       version: freezed == version
+//           ? _value.version
+//           : version // ignore: cast_nullable_to_non_nullable
+//               as String?,
+//     ));
+//   }
+// }
 
-/// @nodoc
-@JsonSerializable()
-class _$_DeviceInfoModel implements _DeviceInfoModel {
-  _$_DeviceInfoModel({this.model, this.id, this.version});
+// /// @nodoc
+// @JsonSerializable()
+// class _$_DeviceInfoModel implements _DeviceInfoModel {
+//   _$_DeviceInfoModel({this.model, this.id, this.version});
 
-  factory _$_DeviceInfoModel.fromJson(Map<String, dynamic> json) =>
-      _$$_DeviceInfoModelFromJson(json);
+//   factory _$_DeviceInfoModel.fromJson(Map<String, dynamic> json) =>
+//       _$$_DeviceInfoModelFromJson(json);
 
-  @override
-  final String? model;
-  @override
-  final String? id;
-  @override
-  final String? version;
+//   @override
+//   final String? model;
+//   @override
+//   final String? id;
+//   @override
+//   final String? version;
 
-  @override
-  String toString() {
-    return 'DeviceInfoModel(model: $model, id: $id, version: $version)';
-  }
+//   @override
+//   String toString() {
+//     return 'DeviceInfoModel(model: $model, id: $id, version: $version)';
+//   }
 
-  @override
-  bool operator ==(dynamic other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _$_DeviceInfoModel &&
-            (identical(other.model, model) || other.model == model) &&
-            (identical(other.id, id) || other.id == id) &&
-            (identical(other.version, version) || other.version == version));
-  }
+//   @override
+//   bool operator ==(dynamic other) {
+//     return identical(this, other) ||
+//         (other.runtimeType == runtimeType &&
+//             other is _$_DeviceInfoModel &&
+//             (identical(other.model, model) || other.model == model) &&
+//             (identical(other.id, id) || other.id == id) &&
+//             (identical(other.version, version) || other.version == version));
+//   }
 
-  @JsonKey(ignore: true)
-  @override
-  int get hashCode => Object.hash(runtimeType, model, id, version);
+//   @JsonKey(ignore: true)
+//   @override
+//   int get hashCode => Object.hash(runtimeType, model, id, version);
 
-  @JsonKey(ignore: true)
-  @override
-  @pragma('vm:prefer-inline')
-  _$$_DeviceInfoModelCopyWith<_$_DeviceInfoModel> get copyWith =>
-      __$$_DeviceInfoModelCopyWithImpl<_$_DeviceInfoModel>(this, _$identity);
+//   @JsonKey(ignore: true)
+//   @override
+//   @pragma('vm:prefer-inline')
+//   _$$_DeviceInfoModelCopyWith<_$_DeviceInfoModel> get copyWith =>
+//       __$$_DeviceInfoModelCopyWithImpl<_$_DeviceInfoModel>(this, _$identity);
 
-  @override
-  Map<String, dynamic> toJson() {
-    return _$$_DeviceInfoModelToJson(
-      this,
-    );
-  }
-}
+//   @override
+//   Map<String, dynamic> toJson() {
+//     return _$$_DeviceInfoModelToJson(
+//       this,
+//     );
+//   }
+// }
 
-abstract class _DeviceInfoModel implements DeviceInfoModel {
-  factory _DeviceInfoModel(
-      {final String? model,
-      final String? id,
-      final String? version}) = _$_DeviceInfoModel;
+// abstract class _DeviceInfoModel implements DeviceInfoModel {
+//   factory _DeviceInfoModel(
+//       {final String? model,
+//       final String? id,
+//       final String? version}) = _$_DeviceInfoModel;
 
-  factory _DeviceInfoModel.fromJson(Map<String, dynamic> json) =
-      _$_DeviceInfoModel.fromJson;
+//   factory _DeviceInfoModel.fromJson(Map<String, dynamic> json) =
+//       _$_DeviceInfoModel.fromJson;
 
-  @override
-  String? get model;
-  @override
-  String? get id;
-  @override
-  String? get version;
-  @override
-  @JsonKey(ignore: true)
-  _$$_DeviceInfoModelCopyWith<_$_DeviceInfoModel> get copyWith =>
-      throw _privateConstructorUsedError;
-}
+//   @override
+//   String? get model;
+//   @override
+//   String? get id;
+//   @override
+//   String? get version;
+//   @override
+//   @JsonKey(ignore: true)
+//   _$$_DeviceInfoModelCopyWith<_$_DeviceInfoModel> get copyWith =>
+//       throw _privateConstructorUsedError;
+// }
