@@ -1,21 +1,21 @@
-// GENERATED CODE - DO NOT MODIFY BY HAND
+// // GENERATED CODE - DO NOT MODIFY BY HAND
 
-part of 'device_info_model.dart';
+// part of 'device_info_model.dart';
 
-// **************************************************************************
-// JsonSerializableGenerator
-// **************************************************************************
+// // **************************************************************************
+// // JsonSerializableGenerator
+// // **************************************************************************
 
-_$_DeviceInfoModel _$$_DeviceInfoModelFromJson(Map<String, dynamic> json) =>
-    _$_DeviceInfoModel(
-      model: json['model'] as String?,
-      id: json['id'] as String?,
-      version: json['version'] as String?,
-    );
+// _$_DeviceInfoModel _$$_DeviceInfoModelFromJson(Map<String, dynamic> json) =>
+//     _$_DeviceInfoModel(
+//       model: json['model'] as String?,
+//       id: json['id'] as String?,
+//       version: json['version'] as String?,
+//     );
 
-Map<String, dynamic> _$$_DeviceInfoModelToJson(_$_DeviceInfoModel instance) =>
-    <String, dynamic>{
-      'model': instance.model,
-      'id': instance.id,
-      'version': instance.version,
-    };
+// Map<String, dynamic> _$$_DeviceInfoModelToJson(_$_DeviceInfoModel instance) =>
+//     <String, dynamic>{
+//       'model': instance.model,
+//       'id': instance.id,
+//       'version': instance.version,
+//     };
