@@ -1,30 +1,21 @@
-// This is a basic Flutter widget test.
-//
-// To perform an interaction with a widget in your test, use the WidgetTester
-// utility in the flutter_test package. For example, you can send tap and scroll
-// gestures. You can also use WidgetTester to find child widgets in the widget
-// tree, read text, and verify that the values of widget properties are correct.
-
-import 'package:flutter/material.dart';
-import 'package:flutter_test/flutter_test.dart';
-
-import 'package:absensi_mahardika/main.dart';
+int findFirstGreater(List<int> array, int threshold) {
+  for (int index = 0; index < array.length; index++) {
+    int value = array[index];
+    if (value > threshold) {
+      return index;
+    }
+    print('Checking element $value at index $index');
+  }
+  return -1;
+}
 
 void main() {
-  testWidgets('Counter increments smoke test', (WidgetTester tester) async {
-    // Build our app and trigger a frame.
-    await tester.pumpWidget(const MyApp());
-
-    // Verify that our counter starts at 0.
-    expect(find.text('0'), findsOneWidget);
-    expect(find.text('1'), findsNothing);
-
-    // Tap the '+' icon and trigger a frame.
-    await tester.tap(find.byIcon(Icons.add));
-    await tester.pump();
-
-    // Verify that our counter has incremented.
-    expect(find.text('0'), findsNothing);
-    expect(find.text('1'), findsOneWidget);
-  });
+  List<int> data = [5, 3, 1, 8, 7, 2];
+  int threshold = 6;
+  int result = findFirstGreater(data, threshold);
+  if (result != -1) {
+    print('The first element greater than $threshold is at index $result.');
+  } else {
+    print('No element greater than the threshold was found.');
+  }
 }
